requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
chardet==4.0.0
colorama==0.4.4
halo==0.0.31
idna==3.2
log-symbols==0.0.14
multidict==5.1.0
pyserial==3.5
python-weather==0.3.6
six==1.16.0
spinners==0.0.24
termcolor==1.1.0
typing-extensions==3.10.0.0
xmltodict==0.12.0
yarl==1.6.3
